refactor(archtest): TAGGROUP-LOOP-FORBIDS-RUNTYPED-01 迁移到 module-path-agnostic [#1302]
移到非 test .go、const 经 PlatformModulePath 派生、dogfood 单源、baseline 缩 1；行为零变（precision-gate + ratchet GREEN）。

- 新增 taggroup_loop_no_runtyped_invariants.go：规则逻辑（consts、scan helpers、CheckTagGroupLoopForbidsRunTyped）从 _test.go 移入非 test 文件
- taggroupLoopArchtestPkg / taggroupLoopTypesevalPkg 由裸字面量改为 PlatformModulePath + "..." 派生形式
- collectKnownTagsBoundObjects 拆出 collectTagsBoundFromAssignStmts / collectTagsBoundFromValueSpecs / taggroupBindIdent 三个顶层 helper，认知复杂度降至 ≤ 15（lint clean）
- taggroup_loop_no_runtyped_test.go 精简为 dogfood 单行（TestTagGroupLoopForbidsRunTyped → Report + CheckTagGroupLoopForbidsRunTyped）+ 精度门 Test_TaggroupLoopFixturePrecisionGate（复用 scanFileForTaggroupViolations 单源）
- testdata/module_path_funnel.baseline 删除 taggroup_loop_no_runtyped_test.go（ratchet-down：baseline 由 83 → 82 条）
